Remove literal formatting from XML definitions
git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@698 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ test/workbench/TestUtil.java
@@ -94,6 +94,8 @@ public class TestUtil
 				deleteFiles(files[i]);
 			}
 			if (files[i].getName().equals("workbench.settings")) continue;
+			if (files[i].getName().equals("workbench.log")) continue;
+			
 			if (!files[i].delete())
 			{
 				System.out.println("Could not delete file: " + files[i].getAbsolutePath());
